Fix support for megaTinyCore, add DxCore support
Testing for the define from build.board is not a good test (I recently updated them to refer to the specific part, instead of just the number of pins, for example) - instead, test for the MEGATINYCORE define, which is defined by Arduino.h and will always be there , and will be for all future versions of megaTinyCore. Also added a test for DxCore, which supports the new AVR Dx-series parts, and does SPI bit order in the same way. I confirmed that the examples now compile again on megaTinyCore with this change - and now compile for DxCore as well