combined read/write methods for an improvement in performance
If the code doesn't run on an 8bit-AVR, the data is prepared by combining
the reads and writes into one buffer, which is then transfered as one
transaction on the bus. A constant decides if the buffer is allocated
on the stack or from the heap, further reducing the overhead for small
reads and writes. This is a tradeof between RAM and CPU: on the 'bigger'
arduino platforms the CPU is a lot faster than the SPI, so using more
RAM/CPU cycles and then letting the DMA do its work is the way to go.

I tested it on an ESP32: on a 10Mhz bus: The overhead for an transaction
was about 0.8us/1.0us (begin/end), by combining reads and writes into a
newly allocated buffer it rises to 1.0us/1.0us (begin/end). But as most of
the typical transactions are combined like getting the register contents
from a chip (1 byte to send, multiple to read), combining cuts out 1.8us
but adds 0.2us, which results in a net gain of 1.6us per transaction. All
in all this results in a gain in typical usecases of around 15%.