beginTransactionWithAssertingCS() and endTransactionWithDeassertingCS()
These two methods are similar to beginTransaction() and endTransaction(),
but, as the name implies, with CS management. They are both public, as
beginTransaction() and endTransaction() are too.
I also moved setChipSelect() further up, so the code is more clearly
structured: the callees are in directly in front of the caller.