write(): removed special case for ESP32 and reuse existing code
There is already a method to send chunks of data: transfer(). Also sending
data byte for byte is really inefficient on platforms with DMA and/or the
ability to send them as buffer/len combo. Using tranfer() not only keeps
all the special casing/logic for sending/receiving in one place, it also is
about 50% faster. In the case if only buffer without prefix_buffer is sent,
there you can use write_and_read() directly.