write_then_read(): removed slow bytewise tranfer
Bytewise sending is really slow on most platforms, as they are optimized
for chunked transfers. I also removed a special case for ESP32, this is
not needed, transfer() is already optimized as it is. The whole transfer is
moved in front of te debug stuff, this should keep the bus less occupied in
case it is activated.