combined read/write methods for an improvement in performance
If the code doesn't run on an 8bit-AVR, the data is prepared by combining
the reads and writes into one buffer, which is then transfered as one
transaction on the bus. A constant decides if the buffer is allocated
on the stack or from the heap, further reducing the overhead for small
reads and writes by reducing heap thrashing. This is a tradeof between
RAM/CPU and bus speed: on the 'bigger' arduino platforms the CPU is a
lot faster than the SPI and there is a lot of RAM avaible, so using more
RAM/CPU cycles and then letting the DMA do its work is the way to go.

This also allows for the use of chunked transfer() in all cases, which
gives an improvement of about 50% over bytewise writes. This is for all
supported platforms except 8bit AVRs, which are specifically #if'd
out. The special case for the ESP32 is therefore removed and ARM M0/M4,
ESP8266, teensy etc should profit of this improvement too, without special
casing each platform by using non standard arduino core extensions.

All in all, this results in a performance improvement for 32bit platforms
of about 65% over bytewise transfers in typical usecases (tested with
an ESP32 with Adafruit_BusIO_Register).