chunked transfers for a big improvement in performance
If the code doesn't run on an 8bit-AVR, the data is prepared by
transmitting the reads and writes in chunks instead of byte for byte. A
constant is used for the chunk size.

This is a tradeof between RAM/CPU and bus speed: on the 'bigger' arduino
platforms the CPU is a lot faster than the SPI and there is a lot of
RAM avaible, so using more RAM/CPU cycles and then letting the DMA do
its work is the way to go.

The chunked transfers also combine the reads and writes, so the dead
time in between is removed, which is especially important for register
reads of SPI-attached chips.

Chunked transfers give an improvement of about 40% over bytewise
ones, additionally +5% in the case of small reads/writes as used in
BusIO_Register by removing the dead time between writing and reading. This
is for all supported platforms except 8bit AVRs, which are specifically
#if'd out. The special case for the ESP32 is therefore removed and
ARM M0/M4, ESP8266, teensy etc should profit of this improvement too,
without special casing each platform by using non standard arduino
core extensions.